services/bot-api/requirements.txt
aiogram==3.4.1
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1
pydantic>=2.4.1,<2.6
pydantic-settings==2.1.0
python-dotenv==1.0.1
google-cloud-secret-manager==2.18.2
google-cloud-aiplatform==1.43.0
asyncpg==0.29.0
httpx==0.27.0
python-json-logger==2.0.7
structlog==24.1.0
